requirements.txt
python-docx
pywin32
pygments
pillow 
html2image
#marko.parse
#evernote
#llama-index